Complete planarrangement feature

+++ core/data/src/main/java/com/nrr/data/repository/RoomTaskRepository.kt
@@ -2,10 +2,10 @@ package com.nrr.data.repository
 
 import com.nrr.database.dao.ActiveTaskDao
 import com.nrr.database.dao.TaskDao
+import com.nrr.database.model.ActiveTask
 import com.nrr.database.model.TaskWithStatus
 import com.nrr.database.model.asEntity
 import com.nrr.database.model.asExternalModel
-import com.nrr.model.ActiveStatus
 import com.nrr.model.Task
 import com.nrr.model.TaskPeriod
 import kotlinx.coroutines.flow.Flow
@@ -16,14 +16,14 @@ internal class RoomTaskRepository @Inject constructor(
     private val taskDao: TaskDao,
     private val activeTaskDao: ActiveTaskDao
 ) : TaskRepository {
-    override fun getAllTasks(): Flow<List<Task>> =
+    override fun getTasks(): Flow<List<Task>> =
         taskDao.getAllTasks().map {
             it.map(TaskWithStatus::asExternalModel)
         }
 
-    override fun getAllActiveTasksByPeriod(period: TaskPeriod): Flow<List<Task>> =
+    override fun getActiveTasksByPeriod(period: TaskPeriod): Flow<List<Task>> =
         activeTaskDao.getAllByPeriod(period).map {
-            it.map(TaskWithStatus::asExternalModel)
+            it.map(ActiveTask::asExternalModel)
         }
 
     override fun getByTitle(title: String): Flow<List<Task>> =
@@ -36,28 +36,41 @@ internal class RoomTaskRepository @Inject constructor(
             it.map(TaskWithStatus::asExternalModel)
         }
 
-    override suspend fun saveTasks(
-        tasks: List<Task>,
-        activeStatus: List<ActiveStatus?>
-    ): List<Long> {
+    override fun getActiveTasksByIds(activeTaskIds: List<Long>): Flow<List<Task>> =
+        activeTaskDao.getAllByIds(activeTaskIds).map {
+            it.map(ActiveTask::asExternalModel)
+        }
+
+    override suspend fun saveTasks(tasks: List<Task>): List<Long> {
         val ids = taskDao.insertTasks(tasks.map(Task::asEntity))
-        activeTaskDao.insertActiveTasks(
-            tasks
-                .filter { it.activeStatus != null }
-                .map { it.activeStatus!!.asEntity(it.id) }
-        )
+        tasks
+            .filter { it.activeStatuses.isNotEmpty() }
+            .flatMap { it.activeStatuses.map { s -> s.asEntity(it.id) } }
+            .run {
+                if (isNotEmpty()) activeTaskDao.insertActiveTasks(this)
+            }
         return ids
     }
 
-    override suspend fun deleteActiveTasks(task: List<Task>): Int =
-        activeTaskDao.deleteActiveTasks(
-            task
-                .filter { it.activeStatus != null }
-                .map { it.activeStatus!!.asEntity(it.id).id }
+    override suspend fun saveActiveTasks(tasks: List<Task>): List<Long> =
+        activeTaskDao.insertActiveTasks(
+            activeTasks = tasks.flatMap {
+                it.activeStatuses.map { s ->
+                    s.asEntity(taskId = it.id)
+                }
+            }
         )
 
+    override suspend fun deleteActiveTasks(task: List<Task>): Int =
+        task
+            .filter { it.activeStatuses.isNotEmpty() }
+            .flatMap { it.activeStatuses.map { s -> s.asEntity(it.id).id } }
+            .run {
+                if (isNotEmpty()) activeTaskDao.deleteActiveTasks(this) else 0
+            }
+
     override suspend fun setActiveTaskAsCompleted(task: Task): Long =
-        task.activeStatus?.let {
+        task.activeStatuses.firstOrNull()?.let {
             activeTaskDao.insertActiveTasks(
                 listOf(it.copy(isCompleted = true).asEntity(task.id))
             ).firstOrNull() ?: 0
